Added my Netbean project. Added my assets.
Added my "Test" project. The project contains 2 packages: the Game package and the Sandbox package. The Game package is a working and tested version of the game. As for the Sandbox package, it's a package that contains test classes and it's generally a enviroment where everyone can do anything without fear of corrupting and/or changing the working code.
I also added and extra assets folder containing some experimented sound effects.