A late Saturday update! Refactored a lot of lines of code. Added some documentation.
Saturday was really productive. I had a problem with the player icons when the game started. The problem was that the game panel was created on startup, just like the menu panel and the tutorial panel. This made the player color selection impossible, because the selections are made on runtime!

My approach:
By having a config file, so that the game can read and write from! The config file can write the color selection file path of the players.
After that, we create a reference to a game panel when the "Start game" button is pressed. Now the constructor of the game panel can read the config file and have the correctly chosen player colors, that were selected from the menu screen. The same idea is applied on the restart button of the options panel. The restart button de-refrences the game panel, affectively deleting it from the program (on runtime). So the game panel is created on runtime and deleted on "restart".

Not the best idea ever but it works just fine with little to no performance hit on RAM usage.
Moral of the story: Games need/have to control memory :')